Add createrenditions command
This commited adds a Django command for generating various file
representations, such as thumbnails, for media objects and is useful
for batch re-generating all representation when e.g. the specs changes.

Preflighted successfully on 50 000 image objects with the current setup
of 3 representations (small, medium, large) per object, so this should
be sufficient for current and any near-future use cases.