Register model signal receivers properly
Since the media app is designed to use polymorphism for models on a
media-type basis we end up with multiple proxy-models (speaking in terms
of Django), and when connecting model signal receivers this needs to be
done for every proxy-model being the sender.

This commit adds a ready method in MediaConfig that connects receivers
for every known proxy-model as the sender, rather than using the
@receiver decorator which is limited to manualy specifying only one
sender.