Improve taxon selection when adding/editing media
Replace the browser-native select widget with the enhanced JavaScript
based select2 widget. This is the same as used in the built-in
AutocompleteSelect Django widget.

This provides a better UX for contributors of media when dealing with
the task of selecting the right related taxon from a large list of
options, that now can be narrowed down by entering a search term.