Add contributor as an attribute for media
The reasoning for this is that a contributor may, for example, have sent
an image to an administrator and asked the administrator to add the
image. Thus, neither the `created_by` field or `photographer_artist`
attribute is sufficient for the use-case of connecting a media entry to
a specific contributor that may or may not be a contributing member of
the application.