Update project dependencies
The latest Django LTS 4.2 branch is now being used and is also the one
that should be used for the upcoming stable release of this project.

Also, psycopg 3 is now being used since support for version 3 was added
in Django 4.2 and psycopg 2 is going to be deprecated for use with
Django later on.