Fix: Initial taxa migration ordering option
Add missing ordering option that somehow did not make it in to the
initial migration previously.

Note: Migrations are not really intended to be modified in this way but
has been and will continue to be until the first stable release. After
the first stable release migrations will be handled properly.