Improve media file storage mechanisms
This commit adds the facilities necessary to distinguish the
implementation of storage between original files and generated versions
of the original files. The initial implementation is pretty much
identical to the previously used default storage, with the primarly
modification being that original files now are stored in a sub directory
called `origins`.