Re-define taxon primary key field
The id for a taxon is not intended to be auto generated, hence the
replacement of AutoField to PositiveBigIntegerField. Even though the ids
has not been autogenerated during an import, this makes it more clear.

Also, an autogenerated id for facts is not necesssary since it is a
one-to-one relationship to taxon and therefore it makes sense to use the
taxon id as primary key and to remove the autogenerated id field.