Replace auto_now with default in Media model
Since `auto_now` and `auto_now_add` always
overrides values passed to the save method it is
impossible to explicity set values for such fields
from the outside.

The fields `created_at` and `updated_at` are now
using a default of `timezone.now` instead of
`auto_now` and `auto_now_add` in order to fix
this issue.