Add manage_others media permission
This commit adds a new permission to the media
model which is intended to be assigned to groups
that can change and delete other users media.

The requirement to be superuser in order to
perform such actions in the admin interface is
now replaced with the requirement to have the
mentioned permission.