Add label for external links
The `label` key/value pair is added to the
structure for every external link. This enables
for more advanced combinations than just
`provider` and `external_id` and makes it easier
to control a good display name for the link in a
more dynamic manner.

Part of issue #15, as a preparation.